bayesian optimization working example using cartpole printing best
